requirements.txt
requests>=2.31.0
feedparser>=6.0.10
python-dateutil>=2.8.2
colorama>=0.4.6
click>=8.1.7